+++ Group/group000.xml
@@ -3,17 +3,6 @@
     <groupID>group000</groupID>
     <Users-List>
         <user>ff</user>
-        <user>qq</user>
     </Users-List>
     <Title-group>Main chat</Title-group>
-    <History-Message>
-        <login>ff</login>
-        <message>d</message>
-        <time>2019-03-24T03:00:52.870+02:00</time>
-    </History-Message>
-    <History-Message>
-        <login>qq</login>
-        <message>к</message>
-        <time>2019-03-24T03:01:48.632+02:00</time>
-    </History-Message>
 </group>
